[Fix] File/task name generation collisions
Tasks that recreate themselves were generating a filename by hashing the file date, but if the data is the same, then the hash is the same, and the new file overrides the existing one right before it's processed/deleted.

Filenames now have a unique id appended to them to prevent these naming collisions.